video:tegra: fix ioctl for 32/64bit compatibility
Changes "long int" to uint in order to ensure
AARCH32/AARCH64 compatibility.

Bug 1466413

Reviewed-on: http://git-master/r/390364
Tested-by: Thomas Cherry <tcherry@nvidia.com>

+++ include/video/tegra_camera.h
@@ -36,7 +36,7 @@ enum {
 struct tegra_camera_clk_info {
 	uint id;
 	uint clk_id;
-	unsigned long rate;
+	uint rate;
 	uint flag;	/* to inform if any special bits need to enabled/disabled */
 };
